+++ CHANGES.md
@@ -9,6 +9,8 @@
 * Modified conversion from JSON to reject zero- and one-dimensional positions.
   * PR: <https://github.com/georust/geojson/pull/225>
 
+* Added `geojson::{to_string, to_string_pretty}` as convenience wrappers around the same `serde_json` methods.
+
 ## 0.24.0
 
 * Added `geojson::{ser, de}` helpers to convert your custom struct to and from GeoJSON.